requirements.txt
# Core PDF processing
PyMuPDF>=1.23.0
reportlab>=4.0.0
pillow>=10.0.0

# OCR and Layout
pytesseract>=0.3.10
tabula-py>=2.9.0
pdf2image>=1.16.3

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
scikit-learn>=1.3.0
openpyxl>=3.1.0

# Validation and comparison
opencv-python>=4.8.0
scikit-image>=0.21.0
imagehash>=4.3.1

# LLM integration (optional - can use API keys)
openai>=1.3.0
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
typing-extensions>=4.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

